Ajustes da altura da nuvem e da velocidade em que ela rega as plantas.

+++ Assets/Scripts/Elements/Plants/Base/IPlant.cs
@@ -7,7 +7,7 @@ public interface IPlant
     GameObject SeedGameObject { get; set; }
     GameObject SproutGameObject { get; set; }
     GameObject TreeGameObject { get; set; }
-    int WaterLevel { get; set; }
+    float WaterLevel { get; set; }
     int WaterLevelMax { get; set; }
     int SproutWaterLevel { get; set; }
     int TreeWaterLevel { get; set; }
